SCE-395 - float to int

+++ pkg/experiment/sensitivity/metadata/metadata.go
@@ -43,8 +43,8 @@ type Aggressor struct {
 
 // Measurement represents metadata of single measurement in sensitivity profile experiment and is agnostic to data store.
 type Measurement struct {
-	Load         float64
-	LoadPointQPS float64
+	Load         int
+	LoadPointQPS int
 	LGParameters string
 }
